commit patches by Vadim & Malcolm:     - by Vadim Catana:       - fixed a couple of compiler warnings for gcc4 on x86-64       - the video-vidix.c and video-dfb.c both contain code         that converts pictures from YV12 format to YUY2. I moved         this common code to a function in utils.c .       - Some channels display white dots at the top and bottom of         the screen due to WSS data encoded in the picture.         This problem has been mentioned on vdr, linux-dvb and         dxr3 mailinglists a few times. I added two new items in         the softdevice menu that allow to cut off a number of lines         from top and bottom of the picture. It is implemented now         only for -vo vidix and dfb with i420/YV12 format and hardware         alpha blending.     - by Malcolm Caldwell: fix for vdr segfaulting

+++ softdevice/mpeg2decoder.c
@@ -3,7 +3,7 @@
  *
  * See the README file for copyright information and how to reach the author.
  *
- * $Id: mpeg2decoder.c,v 1.43 2005/07/03 15:58:30 wachm Exp $
+ * $Id: mpeg2decoder.c,v 1.44 2005/07/15 20:42:16 lucke Exp $
  */
 
 #include <math.h>
@@ -1057,7 +1057,7 @@ start:
 };
 
 int cMpeg2Decoder::seek(offset_t offset, int whence) {
-   printf("unimplemented: seek offset %lld whence %d\n",offset,whence);
+   printf("unimplemented: seek offset %lld whence %d\n", (long long int)offset, whence);
    return -EINVAL;
 };
